pyproject.toml
[tool.poetry]
name = "template_for_ai_pytorch_lightning"
version = "0.1.0"
description = "This is the template for AI-Machine learning task for JX_press member"
authors = ["Yongtae Hwang"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
pandas = "^1.2.4"
torch = "1.8.1"
pytorch-lightning = "^1.3.6"
scikit-learn = "^0.24.2"
albumentations = "^1.0.0"
numpy = "^1.20.3"
torchmetrics = "^0.3.2"
transformers = "^4.7.0"
efficientnet-pytorch = "^0.7.1"
torchvision = "0.9.1"
matplotlib = "^3.4.2"
tweepy = "^3.10.0"
jupyter = "^1.0.0"
mojimoji = "^0.0.11"
emoji = "^1.2.0"
neologdn = "^0.5.1"
pydata-google-auth = "^1.2.0"
google-cloud-datastore = "^2.1.3"
bs4 = "^0.0.1"
fugashi = "^1.1.1"
ipadic = "^1.0.0"
nest-asyncio = "^1.5.1"
twint = {git = "https://github.com/twintproject/twint.git", rev = "origin/master"}
mlflow = "^1.19.0"
onnx = "^1.10.1"
onnxruntime = "^1.8.1"
coloredlogs = "^15.0.1"
sympy = "^1.8"
hydra-colorlog = "^1.1.0"
hydra-core = "^1.1.0"
omegaconf = "^2.1.0"
google-cloud-bigquery = "^2.28.1"
hydra-optuna-sweeper = "^1.1.1"
seaborn = "^0.11.2"
pyarrow = "^5.0.0"
pickle5 = "^0.0.11"
gcsfs = "^2021.10.1"
plotly = "^5.3.1"
lightgbm = "^3.3.1"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]
pysen = {version = "0.9.1", extras = ["lint"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py38"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]